requirements.txt
PyQt5
PyQtWebEngine
qtwidgets
pywebview
tkhtmlview